refactor: move firebase admin logic to api routes
